Add files via upload
A Screenshot of the dashboard which visualizes the survey data of the data professional's currently working in the job market. 
The country of the participants is visualized using a Treemap, which gives the location of the participants.
The chart adjacent to it shows the Average Salary according to the title. It can be filtered for a specific country by clicking in the first visualization.
There are two counts displayed at the top right corner which are for the number of Survey Takers and their average age.
The chart below it shows average number of people that are happy with their salary and their work/life balance ratings.
The Donut chart shows how difficult they thought getting into data was according to their respective roles.
The chart in the bottom left shows what are their favorite  programming languages, where Python turned out to be on the top, followed by R.